feat(conference): integrate ML audio analysis service
- Migrate audio transcription and hold detection to `ConferenceV2/app/services/audio`
- Switch to `AsyncOpenAI` for non-blocking execution
- Add 8-second buffering and timestamped segment extraction
- Integrate with `websocket.py` router
- Clean up legacy `ml-audio-service`
