+++ ConferenceV2/pyproject.toml
@@ -28,6 +28,9 @@ dependencies = [
     "httpx",
     "motor",
     "pymongo",
+    "openai",
+    "numpy",
+    "scipy"
 ]
 
 
@@ -41,4 +44,3 @@ pytest-asyncio = "*"
 pytest-cov = "*"
 [tool.poetry.group.dev.dependencies]
 pytest-cov = "^7.0.0"
-
