--- ml-audio-service/requirements.txt
@@ -1,7 +1,0 @@
-fastapi
-uvicorn
-websockets
-openai
-numpy
-scipy
-python-dotenv
